# docs/openapi.yaml
openapi: 3.0.0
info:
  title: MiniBlog API
  version: 1.0.0
  description: API REST para gestionar autores y posts

servers:
  - url: http://localhost:3000
tags:
  - name: Authors
  - name: Posts

components:
  schemas:

    Author:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        bio:
          type: string
        created_at:
          type: string

    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        published:
          type: boolean
        author_id:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string

paths:

  /authors:
    get:
      tags: [Authors]
      summary: Obtener todos los autores
      responses:
        "200":
          description: Lista de autores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Author"

    post:
      tags: [Authors]
      summary: Crear un autor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                email:
                  type: string
                bio:
                  type: string
      responses:
        "201":
          description: Autor creado
        "400":
          description: Error de validación

  /authors/{id}:
    get:
      tags: [Authors]
      summary: Obtener autor por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Autor encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"
        "404":
          description: No encontrado

    put:
      tags: [Authors]
      summary: Actualizar autor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Author"
      responses:
        "200":
          description: Autor actualizado
        "404":
          description: No encontrado

    delete:
      tags: [Authors]
      summary: Eliminar autor
      responses:
        "200":
          description: Autor eliminado
        "404":
          description: No encontrado

  /posts:
    get:
      tags: [Posts]
      summary: Obtener todos los posts
      responses:
        "200":
          description: Lista de posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

    post:
      tags: [Posts]
      summary: Crear post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content, author_id]
              properties:
                title:
                  type: string
                content:
                  type: string
                author_id:
                  type: integer
                published:
                  type: boolean
      responses:
        "201":
          description: Post creado
        "400":
          description: Error de validación

  /posts/{id}:
    get:
      tags: [Posts]
      summary: Obtener post por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: No encontrado

    put:
      tags: [Posts]
      summary: Actualizar post
      responses:
        "200":
          description: Post actualizado
        "404":
          description: No encontrado

    delete:
      tags: [Posts]
      summary: Eliminar post
      responses:
        "200":
          description: Post eliminado
        "404":
          description: No encontrado

  /posts/author/{authorId}:
    get:
      tags: [Posts]
      summary: Obtener posts por autor
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Posts del autor